Use Xorshift64* to generate noise with less artifacts
Xor128 had too large state vector that is sensitive to initial state.
Namely, seed value only affects the first word of the state. Shuffling
the value across all state bits takes CPU time and we don't want that
for ever vertex in the grid. Xorshift64 has just 1 u64 as the whole
state and less sensitive to initial distribution of bits.