Try to render train tracks without a gap at nodes
When the zoom level is high enough to render track in details, we need
to compute the parallel offset to render 2 track rails, but we didn't
render the last segment since the tangent vector was undefined.
We simply use the last segment's normal vector, which may not be correct
but better than nothing.

We also fix the color to render ghost in simple mode when attempting to
build a rail in water.