Allow extending a track to either forward or backward
Previously, you could only extend forward. The goal is that the user
can't tell the difference between forward and backward, because train
tracks don't have one.

We use a similar UI to select a direction to extend to Factorio. If the
user clicks a point closer to a segment, it will extend to the other
side of the segment. It is indicated by a yellow triangle.