Re-interpolate nodes to make uniform interval
Since Bezier curves don't have a fixed form formula for length along
the curve, we need to approximate it by sampling. Technically, Other
segment shapes (Line and Arc) don't need sampling, because they have
closed form formula for the length, but we treat them the same way as
the Bezier.