Maintain train speed in each car
Because each car of a train can be on a different path, we can't use a
single speed value for all the cars. Paths can have opposite direction
from connected ones, so we need to switch sign of the speed for each
path.

That said, we don't want to maintain different speed values for the
whole train, because they should be the same for all the cars.
The approach we take to solve this issue is to have separate
"adjustment" step after the speed and positions are calculated, so
the car interval and speeds are consistent. This adjustment is done
pairwise, so it is not a global relaxation for more than 2 cars.
This introduces a "springy" effect at the end of paths, which should be
ideally "stiff".
This adjustment is tricky among different paths, so we implement it for
cars in the same path in this commit.

We also show a debug print on UI frame to show each car's path id, s
value and speed.

We also prevent the user from switching the path switch if the train is
on a switch point, because it would mess up the connections between
cars.