Add s_length property to capture fractional parts of the length
Previously, we assumed that a path lengh is a multiple of
SEGMENT_LENGTH, which is obviously not true in general. It caused weird
behaviors in transitioning paths. Now that we use s_length, which is a
f64, to calculate transition, we do not suffer from loss of precision,
making the transition much smoother.