Add VarDef statement to define variables and constants
Constants are determined by if the sub-expression contains a reference
to "x". It is similar to GLSL's position argument, which is given by
the evaluator. We perform constant-folding to allow variables in the
arguments of perlin_noise or white_noise. In this way, we can look up
the noise based on the result of another noise.

There is another notable issue in the parser that the syntax can be
ambiguous if the first token was an indentifier. namely, "a(0)" and "a +
1" are expressions, but "a = 0;" is a variable definition statement,
which is not decidable at the point "a" is parsed. Recursive descent
parsers can resolve the ambiguity by backtracking, but it can make
inefficiency by re-evaluating the same sub-expression multiple times.
Therefore, I used partial application of parsers to avoid backtracking.