Use one margin pixel to make the tiles seamless
Now the gradient can use the next pixel at the border so it can shade
the pixel correctly.

I was a bit worried that having an odd number of pixels may round up to
multiple of words, so far it's working. Memory may be padded in an
inefficient way though.