Offload heightmap generation to another thread
Unfortunately rayon did not provide us with the feature we need, which
is an asynchronous task queue.

We implemented one with standard library threads and mpsc, but it can
only support 1 worker. mpmc is not stabilized yet.

It also fails in Wasm so we need a way to switch this feature in wasm
build or use SharedMemoryBuffer and multiple Wasm workers, but it seems
not quite easy ride.